Flicker demonstration
The sample project seems to not flicker too much when zoomed in a lot, but when you zoom out to fill the entire user path it demonstrates the flicker easily. Even though only one annotation is being added / removed, the main user route annotation flickers and it has to be redrawn.